ADOP-1898: Check and send order changes
