+++ src/main/java/uk/gov/hmcts/reform/adoption/adoptioncase/common/CommonPageBuilder.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.page;
+package uk.gov.hmcts.reform.adoption.adoptioncase.common;
 
 import org.apache.commons.collections4.CollectionUtils;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
@@ -6,12 +6,7 @@ import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.ccd.sdk.type.DynamicList;
 import uk.gov.hmcts.ccd.sdk.type.DynamicListElement;
 import uk.gov.hmcts.reform.adoption.adoptioncase.common.CaseEventCommonMethods;
-import uk.gov.hmcts.reform.adoption.adoptioncase.common.CommonPageBuilder;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.SelectedMessage;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.MessageSendDetails;
-import uk.gov.hmcts.reform.adoption.common.ccd.CcdPageConfiguration;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.*;
 import uk.gov.hmcts.reform.adoption.common.ccd.PageBuilder;
 
 import java.util.ArrayList;
@@ -19,20 +14,22 @@ import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
 
-public class SendOrReply implements CcdPageConfiguration {
-    @Override
-    public void addTo(PageBuilder pageBuilder) {
-        CommonPageBuilder.sendOrReplyCommonPage(pageBuilder, "");
-//        pageBuilder
-//            .page("pageSendOrReply1", this::midEvent)
-//            .mandatory(CaseData::getMessageAction)
-//            .mandatory(CaseData::getReplyMsgDynamicList, "messageAction=\"replyMessage\"");
-//        replyMessageBuilder(pageBuilder, "messageAction=\"replyMessage\"");
-//        messageBuilder(pageBuilder, "messageAction=\"sendMessage\" OR replyMessage=\"Yes\"");
+public class CommonPageBuilder {
 
+    private CommonPageBuilder() {
     }
 
-    public void messageBuilder(PageBuilder pageBuilder,String condition) {
+    public static void sendOrReplyCommonPage(PageBuilder pageBuilder, String type) {
+        pageBuilder
+            .page("pageSendOrReply1", CommonPageBuilder::sendMesgMidEvent)
+            .showCondition(type)
+            .mandatory(CaseData::getMessageAction)
+            .mandatory(CaseData::getReplyMsgDynamicList, "messageAction=\"replyMessage\"");
+        replyMessageBuilder(pageBuilder, "messageAction=\"replyMessage\"");
+        messageBuilder(pageBuilder, "messageAction=\"sendMessage\" OR replyMessage=\"Yes\"");
+    }
+
+    public static void messageBuilder(PageBuilder pageBuilder,String condition) {
         pageBuilder.page("pageSendOrReply3")
             .showCondition(condition)
             .label("sendMessageLab", "## Send a message","messageAction=\"sendMessage\"")
@@ -50,7 +47,7 @@ public class SendOrReply implements CcdPageConfiguration {
             .done();
     }
 
-    public void replyMessageBuilder(PageBuilder pageBuilder, String condition) {
+    public static void replyMessageBuilder(PageBuilder pageBuilder, String condition) {
         pageBuilder.page("pageSendOrReply2")
             .showCondition(condition)
             .label("labelReplyMes", "## Reply to message")
@@ -64,19 +61,19 @@ public class SendOrReply implements CcdPageConfiguration {
 
     }
 
-    private AboutToStartOrSubmitResponse<CaseData, State> midEvent(CaseDetails<CaseData, State> details,
+    private static AboutToStartOrSubmitResponse<CaseData, State> sendMesgMidEvent(CaseDetails<CaseData, State> details,
                                                                    CaseDetails<CaseData, State> detailsBefore) {
         CaseData caseData = details.getData();
         List<DynamicListElement> listElements = new ArrayList<>();
         CaseEventCommonMethods.prepareDocumentList(caseData).forEach(item -> listElements.add(DynamicListElement.builder()
-                             .label(item.getDocumentLink().getFilename()).code(UUID.fromString(item.getMessageId())).build()));
+                                                                                                  .label(item.getDocumentLink().getFilename()).code(UUID.fromString(item.getMessageId())).build()));
         caseData.setAttachDocumentList(DynamicList.builder().listItems(listElements).value(DynamicListElement.EMPTY).build());
 
         if (CollectionUtils.isNotEmpty(caseData.getListOfOpenMessages()) && caseData.getReplyMsgDynamicList() != null) {
             var messageDetails = new SelectedMessage();
             var selectedObject = caseData.getListOfOpenMessages().stream()
                 .filter(item -> item.getValue().getMessageId().equalsIgnoreCase(caseData.getReplyMsgDynamicList()
-                                                                               .getValueCode().toString())).findFirst();
+                                                                                    .getValueCode().toString())).findFirst();
             messageDetails.setMessageId(selectedObject.get().getId().toString());
             messageDetails.setUrgency(selectedObject.get().getValue().getMessageUrgencyList().getLabel());
             messageDetails.setMessageContent(selectedObject.get().getValue().getMessageText());
@@ -90,7 +87,4 @@ public class SendOrReply implements CcdPageConfiguration {
             .data(caseData)
             .build();
     }
-
-
-
 }
